# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for processing bracketed images with AutoEnhance AI HDR merging. This API handles order creation, image uploads, HDR processing, and real-time status updates via Supabase Realtime.",
        "title": "Instant HDR Backend API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/orders/{order_id}/images": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Returns a list of all processed images for an order with their download status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "List processed images",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID (UUID)",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ImagesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/orders/{order_id}/images/{image_id}": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Deletes a processed image from AutoEnhance AI and all associated files from Supabase Storage and database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Delete a processed image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID (UUID)",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Image ID from AutoEnhance",
                        "name": "image_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/orders/{order_id}/images/{image_id}/download": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Downloads a processed image from AutoEnhance and stores it in Supabase Storage.\n\nQuality Options:\n- \"thumbnail\": 400px width (~50-100KB) - List view\n- \"preview\": 800px width (~150-250KB) - Gallery view (DEFAULT)\n- \"medium\": 1920px width (~500KB-1MB) - Full screen\n- \"high\": Full resolution (~2-5MB) - Client delivery\n- \"custom\": Specify max_width or scale\n\nFormat Options: \"jpeg\" (default), \"png\", \"webp\"\n\nWatermark (defaults to true = FREE):\n- true: FREE download with watermark\n- false: COSTS 1 CREDIT (unwatermarked)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Download processed image to Supabase Storage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID (UUID)",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Image ID from AutoEnhance",
                        "name": "image_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Download options",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DownloadImageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DownloadImageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.DownloadImageRequest": {
            "type": "object",
            "properties": {
                "format": {
                    "description": "Format - Image format: \"jpeg\" (default), \"png\", or \"webp\"\nExample: \"jpeg\"",
                    "type": "string",
                    "example": "jpeg"
                },
                "max_width": {
                    "description": "MaxWidth - Custom width in pixels (only used when quality=\"custom\")\nExample: 2400",
                    "type": "integer",
                    "example": 2400
                },
                "quality": {
                    "description": "Quality preset - Options: \"thumbnail\" (400px), \"preview\" (800px), \"medium\" (1920px), \"high\" (full res), or \"custom\"\nDefault: \"preview\"\nExample: \"high\"",
                    "type": "string",
                    "example": "high"
                },
                "scale": {
                    "description": "Scale - Scale factor (only used when quality=\"custom\")\nExample: 0.5 for 50% of original size",
                    "type": "number",
                    "example": 0.5
                },
                "watermark": {
                    "description": "Watermark - Whether to include watermark. Defaults to true (FREE). Set to false to use 1 credit (unwatermarked)\nExample: true",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.DownloadImageResponse": {
            "type": "object",
            "properties": {
                "credit_used": {
                    "description": "CreditUsed indicates if this download cost a credit",
                    "type": "boolean",
                    "example": false
                },
                "file_size": {
                    "description": "FileSize in bytes",
                    "type": "integer",
                    "example": 2621440
                },
                "format": {
                    "description": "Format of the downloaded image",
                    "type": "string",
                    "example": "jpeg"
                },
                "image_id": {
                    "description": "ImageID from AutoEnhance",
                    "type": "string",
                    "example": "img_abc123"
                },
                "message": {
                    "description": "Message with download details",
                    "type": "string",
                    "example": "Image downloaded successfully (FREE with watermark) - Quality: high, Resolution: full"
                },
                "quality": {
                    "description": "Quality preset used",
                    "type": "string",
                    "example": "high"
                },
                "resolution": {
                    "description": "Resolution achieved (e.g., \"400px\", \"800px\", \"1920px\", \"full\")",
                    "type": "string",
                    "example": "full"
                },
                "url": {
                    "description": "URL to access the image in Supabase Storage (publicly accessible)",
                    "type": "string",
                    "example": "https://project.supabase.co/storage/v1/object/public/hdr-images/users/user123/orders/order456/img_abc123_high.jpg"
                },
                "watermark": {
                    "description": "Watermark indicates if watermark was applied (true = FREE, false = COSTS 1 CREDIT)",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.ImageResponse": {
            "type": "object",
            "properties": {
                "downloaded": {
                    "type": "boolean"
                },
                "enhance_type": {
                    "type": "string"
                },
                "high_res_downloaded": {
                    "type": "boolean"
                },
                "high_res_url": {
                    "description": "Supabase URL for high-res",
                    "type": "string"
                },
                "image_id": {
                    "type": "string"
                },
                "image_name": {
                    "type": "string"
                },
                "preview_downloaded": {
                    "type": "boolean"
                },
                "preview_url": {
                    "description": "Supabase URL for preview",
                    "type": "string"
                },
                "processing_settings": {
                    "type": "object",
                    "additionalProperties": true
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.ImagesResponse": {
            "type": "object",
            "properties": {
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ImageResponse"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}